Separate ArgsToFrontendOptionsConverter, ArgsToFrontendOutputsConveter, InputFile, FrontendInputs, ArgsToFrontendInputsConverter into separate files.
Moving thngs around before putting outputs in inputs.
